% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_preseason.R
\name{add_preseason}
\alias{add_preseason}
\title{Initialize Pre Season Projections into modeling format}
\usage{
add_preseason(run_plays, pass_plays, drives, preseason_vals,
  seed = round(runif(1) * 1e+07))
}
\arguments{
\item{run_plays}{the cleaned run plays after the \code{add_epa} function}

\item{pass_plays}{the cleaned pass plays after the \code{add_epa} function}

\item{drives}{the fixed drives after the \code{fix_drives} function}

\item{preseason_vals}{the output from one of the \code{generate_preseason_} functions}

\item{seed}{the random seed to use. Default is a random seed}
}
\value{
A list containing the initial run, pass, and drive values.
}
\description{
Takes the preseason projections for teams based on the \code{generate_preseason}
function and puts them in the proper format to use to build the team
model estimates. It does this by "building" a game's worth of plays for
each team from their preseason forecasts. See the vignette for more info.
}
\examples{
years <- 2013:2014
plays <- readin("play", years)
teams <- readin("team", years)
runs <- readin("rush", years)
pass <- readin("pass", years)
games <- readin("game", years)
conf <- readin("conference", years)

epa_model <- expected_points_build(plays[plays$Year != 2014, ], drives[drives$Year != 2014, ])
fixed_games <- fix_games(games)
drives <- fix_drives(fixed_games, drives)
model_plays <- combine_run_pass(runs, pass, fixed_games) \%>\% remove_garbage \%>\% fix_fcs(teams, conf) \%>\% add_epa(epa_model)
model_values <- generate_preseason_mlm(run_plays = model_plays[["run_info"]], pass_plays = model_plays[["pass_info"]]) \%>\% add_preseason(run_plays = model_plays[["run_info"]], pass_plays = model_plays[["pass_info"]], drives = drives, preseason_vals = .)
}

