% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expected_points_build.R
\name{expected_points_build}
\alias{expected_points_build}
\title{Builds an Expected Points Model}
\usage{
expected_points_build(plays, drives)
}
\arguments{
\item{plays}{the "play.csv" data frame for the years to use to build the model}

\item{drives}{the "drive.csv" data frame}
}
\value{
a named vector where the names are the yardlines on the field and the values are the estimated point values associated with that spot.
}
\description{
Takes in play by play data and drive info to estimate the value of each spot
on the field. The model is built using a B-spline regression model
predicting the number of points scored on the drives based on the current
field position.
}
\examples{
plays <- readin("play", 2014)
drives <- readin("drive", 2014)
epa_values <- expected_points_build(plays, drives)
\dontrun{ggplot2::qplot(x = as.numeric(names(epa_values)), y = epa_values, xlab = "Distance From End Zone", ylab = "Expected Points", geom = "line", main = "Expected Points Model Results")}
}

