% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fix_drives.R
\name{fix_drives}
\alias{fix_drives}
\title{Adds the week, home team, and away team to each drive info}
\usage{
fix_drives(games, drives)
}
\arguments{
\item{games}{the game information from the \code{fix_games} function}

\item{drives}{the "drive.csv" values from cfb stats data.}
}
\value{
a data frame containing with the fixed drive info
}
\description{
This function merges the fixed game info to the drive data so that we have
the week of each game, home and away team, and who is on offense and defense
}
\examples{
years <- 2014
drives <- readin("drive", years)
games <- readin("game", years)
fixed_drives <- fix_games(games) \%>\% fix_drives(drives)
}

