% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fix_fcs.r
\name{fix_fcs}
\alias{fix_fcs}
\title{Convert all FCS teams to one "team"}
\usage{
fix_fcs(all_info, teams, conf)
}
\arguments{
\item{all_info}{the cleaned play by play data frame}

\item{teams}{the "team.csv" values from cfb stats data.}

\item{conf}{the "conference.csv" values from the cfb stats data.}
}
\value{
a data frame containing with the fixed offense and defense team codes.
}
\description{
FCS teams usually only show up once or twice in each season. To overcome
the issues with the small sample size of predicting these games this
function treats all fcs teams as one team.
}
\examples{
years <- 2014
teams <- readin("team", years)
runs <- readin("rush", years)
pass <- readin("pass", years)
games <- readin("game", years)
conf <- readin("conference", years)
game_info <- fix_games(games)
no_fcs <- combine_run_pass(rush, pass, game_info) \%>\% fix_fcs(teams, conf)
}

