% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fix_games.R
\name{fix_games}
\alias{fix_games}
\title{Adds date and week of season to game data}
\usage{
fix_games(games)
}
\arguments{
\item{games}{the "game.csv" value from cfb stats data.}
}
\value{
a data frame containing the augmented game data
}
\description{
This function finds the week of the season that each game was played.
}
\examples{
games <- readin("game", 2009:2014)
game_info <- fix_games(games)
}

