% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_preseason_mlm.r
\name{generate_preseason_mlm}
\alias{generate_preseason_mlm}
\title{Generates the preseason projections for the teams}
\usage{
generate_preseason_mlm(run_plays, pass_plays)
}
\arguments{
\item{run_plays}{the cleaned run plays after the \code{add_epa} function}

\item{pass_plays}{the cleaned pass plays after the \code{add_epa} function}
}
\value{
a data frame containing the estimated team values at the end of each year.
}
\description{
This function generates preseason projections using mulilevel models to
estimate the contribution of each team's offense and defense provides to
their Expected Points Added per Play.
}
\examples{
years <- 2013:2014
plays <- readin("play", years)
teams <- readin("team", years)
runs <- readin("rush", years)
pass <- readin("pass", years)
games <- readin("game", years)
conf <- readin("conference", years)
epa_model <- expected_points_build(plays[plays$Year != 2014, ], drives[drives$Year != 2014, ])
fixed_games <- fix_games(games)
model_plays <- combine_run_pass(runs, pass, fixed_games) \%>\% remove_garbage \%>\% fix_fcs(teams, conf) \%>\% add_epa(epa_model)
preseason_vals <- generate_preseason_mlm(model_plays[[1]], model_plays[[2]])
}

