% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readin.R
\name{readin}
\alias{readin}
\title{Wrapper for reading in multiple years of CFB Stats data}
\usage{
readin(file, yearvec)
}
\arguments{
\item{file}{The name of the file to read in. The extension `.csv` is unnecessary.}

\item{yearvec}{A vector containing the years to read in data for}
}
\value{
A data frame containing the years of values
}
\description{
This function reads in multiple years of files and returns them all in
one data frame. The new data frame contains an additional column labeling
the year that the observation was read in from. Additionally any spaces ( )
in column names are replaced with periods (.).
}
\examples{
plays <- readin("play", 2009:2012)
}

