% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remove_garbage.R
\name{remove_garbage}
\alias{remove_garbage}
\title{Filters out plays designated as garbage time plays}
\usage{
remove_garbage(plays)
}
\arguments{
\item{plays}{a play by play data frame.}
}
\value{
a data frame containing all non-garbage plays
}
\description{
Garbage time plays are plays in the first quarter with leads greater than
28, 2nd quarter with leads greater than 24, 3rd quarter with leads greater
than 21, and the 4th quarter with leads greater than 16.
}
\examples{
pass <- readin("pass", 2014)
rush <- readin("rush", 2014)
games <- readin("game", 2014)
game_info <- fix_games(games)
no_garbage <- combine_run_pass(rush, pass, game_info) \%>\% remove_garbage
}

