% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/weekly_values_mlm.R
\name{weekly_values_mlm}
\alias{weekly_values_mlm}
\title{Generates weekly team-unit estimates}
\usage{
weekly_values_mlm(model_runs, model_pass, model_drives)
}
\arguments{
\item{model_runs}{the cleaned run plays after the \code{add_epa} function}

\item{model_pass}{the cleaned pass plays after the \code{add_epa} function}

\item{model_drive}{the fixed drives after the \code{fix_drives} function}
}
\value{
A data frame containing the weekly teem estimates
}
\description{
This function takes in the formatted play by play data and build weekly team
estimates to be used in the point prediction models. This function uses
multilevel models on play by play data prior to a given week to build the
estimates going in to that week.
}
\examples{
years <- 2013:2014
plays <- readin("play", years)
teams <- readin("team", years)
runs <- readin("rush", years)
pass <- readin("pass", years)
games <- readin("game", years)
conf <- readin("conference", years)

epa_model <- expected_points_build(plays[plays$Year != 2014, ], drives[drives$Year != 2014, ])
fixed_games <- fix_games(games)
drives <- fix_drives(fixed_games, drives)
model_plays <- combine_run_pass(runs, pass, fixed_games) \%>\% remove_garbage \%>\% fix_fcs(teams, conf) \%>\% add_epa(epa_model)
model_values <- generate_preseason_mlm(run_plays = model_plays[["run_info"]], pass_plays = model_plays[["pass_info"]]) \%>\% add_preseason(run_plays = model_plays[["run_info"]], pass_plays = model_plays[["pass_info"]], drives = drives, preseason_vals = .)
weekly_values <- weekly_values_mlm(model_runs = model_values[["run_data"]], model_pass = model_values[["pass_data"]], model_drives = model_values[["drives_data"]])
}

