% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combine_run_pass.R
\name{combine_run_pass}
\alias{combine_run_pass}
\title{Combines running and passing plays}
\usage{
combine_run_pass(runs, pass)
}
\arguments{
\item{runs}{the "rush.csv" data frame for the years to use to build the model}

\item{pass}{the "pass.csv" data frame}
}
\value{
a data frame containing all rushing and passing plays
}
\description{
The running and passing plays have different formats, so this function puts
them in the same format and binds them into one data frame. In addition this
function turns the rushing plays that are sacks into passing plays.
}
\examples{
pass <- readin("pass", 2014)
rush <- readin("rush", 2014)
all_plays <- combine_run_pass(rush, pass)
}

